PR requirements
