+++ src/model/Session.ts
@@ -47,18 +47,3 @@ export type Play = {
   card: Option<Card>;
   target: Option<Player>;
 };
-
-export type DocumentSnapshot = firebase.firestore.DocumentSnapshot<
-  firebase.firestore.DocumentData
->;
-
-export type QuerySnapshot = firebase.firestore.QuerySnapshot<
-  firebase.firestore.DocumentData
->;
-export type DocumentReference = firebase.firestore.DocumentReference<
-  firebase.firestore.DocumentData
->;
-
-export type CollectionReference = firebase.firestore.CollectionReference<
-  firebase.firestore.DocumentData
->;
