+++ src/store/actions.ts
@@ -2,7 +2,6 @@ import { Dispatch } from "redux";
 import { Player, PlayerWithId } from "../model/Player";
 import { requestCreateSession, requestAddPlayer } from "../api/db/session";
 import { SessionWithId } from "../model/Session";
-import { buildOne } from "../model/Card";
 export const CREATE_SESSION = "CREATE_SESSION" as const;
 export const SET_PLAYER = "SET_PLAYER" as const;
